+++ NEWS.md
@@ -2,8 +2,7 @@
 
 - add new parameter `standalone` to `download_cruncher()`.
 
-- by default, the option `"is_cruncher_v3"` is now set to `TRUE`.
-A startup message is added.
+- A startup message is added indicating which version of the JWSACruncher is used by default (defined by the option `"is_cruncher_v3"`).
 
 - new function `download_jdemetra()` to download the JDemetra+ software.